requirements.txt
numpy==1.26.4
opencv-python==4.11.0.86
keras==3.5.0
pandas==2.2.3
tensorflow==2.17.0
urllib3==2.3.0
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.2